Revert "Move BlobStorageLogWriter to Interpreters"
This reverts commit ca3d80102365e76d931be016638b1ca506dffb86.
