+++ src/IO/WriteBufferFromS3.h
@@ -11,7 +11,7 @@
 #include <IO/WriteSettings.h>
 #include <Storages/StorageS3Settings.h>
 #include <Common/threadPoolCallbackRunner.h>
-#include <Interpreters/BlobStorageLogWriter.h>
+#include <IO/S3/BlobStorageLogWriter.h>
 #include <Common/BufferAllocationPolicy.h>
 
 #include <memory>
